Rename & clean some files
