+++ src/shaders/symbols/glsl_filter.rs
